+++ native_src/src/core/lib/transport/parsed_metadata.h
@@ -138,7 +138,7 @@ class ParsedMetadata {
   // Construct metadata from a string key, slice value pair.
   ParsedMetadata(Slice key, Slice value)
       : vtable_(ParsedMetadata::KeyValueVTable(key.as_string_view())),
-        transport_size_(key.size() + value.size() + 32) {
+        transport_size_((uint32_t)(key.size() + value.size() + 32)) {
     value_.pointer =
         new std::pair<Slice, Slice>(std::move(key), std::move(value));
   }
@@ -178,7 +178,7 @@ class ParsedMetadata {
     result.vtable_ = vtable_;
     result.value_ = value_;
     result.transport_size_ =
-        TransportSize(vtable_->key(value_).length(), value.length());
+        TransportSize((uint32_t)(vtable_->key(value_).length()), (uint32_t)(value.length()));
     vtable_->with_new_value(&value, on_error, &result);
     return result;
   }
